requirements.txt
aiofiles==23.2.1
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==7.1.2
click-completion==0.5.2
colored==1.4.2
configobj==5.0.8
configparser==6.0.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
frontend==0.0.3
future==0.18.3
gitdb==4.0.5
GitPython==3.1.11
h11==0.14.0
httplib2==0.22.0
idna==3.6
imageio==2.33.1
img==2.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==2.11.2
looseversion==1.3.0
lxml==4.9.4
MarkupSafe==1.1.1
merger==0.5
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
Pillow==10.1.0
pillow-heif==0.1.5
prompt-toolkit==3.0.8
prov==2.0.0
pycparser==2.21
pydot==1.4.2
pyfiglet==0.8.post1
pyheif==0.7.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pypdf==3.17.3
python-dateutil==2.8.2
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
reportlab==4.0.8
requests==2.31.0
scipy==1.11.4
sh==1.14.1
shellingham==1.3.2
simplejson==3.19.2
six==1.15.0
smmap==3.0.4
sniffio==1.3.0
starlette==0.34.0
traits==6.3.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.5
